requirements.txt
streamlit
streamlit-chat
python-dotenv
pymongo
langchain
openai